Adding verbose logging and fixed uint16 to string issue

+++ config.go
@@ -16,7 +16,7 @@ type ConnectionConfig struct {
     S3Bucket    string      `json:"s3_bucket"`
     LocalDir    string      `json:"local_directory"`
     Hostname    string      `json:"hostname"`
-    Port        uint16      `json:"port"`
+    Port        int         `json:"port"`
     Username    string      `json:"username"`
     Password    string      `json:"password"`
     Databases   []string    `json:"databases"`
